fix: run security validations on security objects only
